created types file for better organization
